Implement buffer page structure
resolved #2

+++ database/project2/src/index.c
@@ -4,7 +4,7 @@ uint64_t numOfTables = 0;
 
 // return execution number of open_table
 // open first table: return 1 
-int open_table(char *pathname) {
+uint64_t open_table(char *pathname) {
     if (file_open_table(pathname) == 0) { //not exist
         //init table
         page_t* header = (page_t*)malloc(sizeof(struct page_t));
